users-backend/pyproject.toml
[tool.poetry]
name = "users"
version = "0.1.0"
description = ""
authors = ["Marco Acierno <marcoaciernoemail@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
starlette = "^0.14.1"
uvicorn = "^0.13.3"
strawberry-graphql = "0.49.0"
SQLAlchemy = "1.4.2"
alembic = "^1.5.2"
psycopg2 = "^2.8.6"
asyncpg = "^0.21.0"
PyJWT = "^2.0.1"
pydantic = "^1.7.3"
email-validator = "^1.1.2"
argon2-cffi = "^20.1.0"
httpx = "^0.16.1"
Authlib = "^0.15.3"
itsdangerous = "^1.1.0"
mangum = "^0.10.0"
pythonit-toolkit = "0.1.18"

[tool.poetry.dev-dependencies]
pdbpp = "^0.10.2"
ward = "^0.48.0-beta.0"
requests = "^2.25.1"
taskipy = "^1.6.0"
factory-boy = "^3.2.0"
asgi-lifespan = "^1.0.1"
coverage = "^5.4"
time-machine = "^2.0.1"

[tool.taskipy.tasks]
server = "uvicorn main:app --port 8050 --reload"
test = "RUNNING_TESTS=1 python scripts/prepare_db.py && RUNNING_TESTS=1 coverage run -m ward"
"test:coverage" = "task test || coverage xml -i && coverage xml -i"
"test:debug" = "task test --no-capture-output"
migrate = "alembic upgrade head"
makemigrations = "alembic revision --autogenerate -m"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
]

[tool.coverage.run]
branch = true
include = [
    "users/*",
]
